[#906] Rename notifications store for consistency
