+++ tests/unit/vue/composables/useExceptionReporting.spec.ts
@@ -1,16 +1,16 @@
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
-import { useExceptionReporting } from '@/composables/useExceptionReporting'
-import api from '@/utils/api'
+import { useExceptionReporting } from '@/composables/useExceptionReporting';
+import type { ExceptionData } from '@/types/exceptions'; // Add this type
+import api from '@/utils/api';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-// Mock the api module
 vi.mock('@/utils/api', () => ({
   default: {
-    post: vi.fn()
-  }
-}))
+    post: vi.fn(),
+  },
+}));
 
 describe('useExceptionReporting', () => {
-  const mockExceptionData = {
+  const mockExceptionData: ExceptionData = {
     message: 'Test exception',
     type: 'Error',
     stack: 'Error stack trace',
@@ -18,36 +18,37 @@ describe('useExceptionReporting', () => {
     line: 1,
     column: 1,
     environment: 'test',
-    release: '1.0.0'
-  }
+    release: '1.0.0',
+  };
 
-  let consoleErrorMock: ReturnType<typeof vi.spyOn>
+  let consoleErrorSpy: ReturnType<typeof vi.spyOn>;
 
   beforeEach(() => {
-    vi.clearAllMocks()
-    consoleErrorMock = vi.spyOn(console, 'error').mockImplementation(() => {})
-  })
+    vi.clearAllMocks();
+    consoleErrorSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
+  });
 
   afterEach(() => {
-    consoleErrorMock.mockRestore()
-  })
+    consoleErrorSpy.mockRestore();
+  });
 
-  it('should report exceptions correctly', async () => {
-    vi.mocked(api.post).mockResolvedValueOnce({})
-    const { reportException } = useExceptionReporting()
+  it('reports exceptions successfully', async () => {
+    vi.mocked(api.post).mockResolvedValueOnce({});
+    const { reportException } = useExceptionReporting();
 
-    await reportException(mockExceptionData)
+    await reportException(mockExceptionData);
 
-    expect(api.post).toHaveBeenCalledWith('/api/v2/exception', mockExceptionData)
-    expect(api.post).toHaveBeenCalledTimes(1)
-  })
+    expect(api.post).toHaveBeenCalledWith('/api/v2/exception', mockExceptionData);
+    expect(api.post).toHaveBeenCalledTimes(1);
+  });
 
-  it('should handle exceptions gracefully', async () => {
-    vi.mocked(api.post).mockRejectedValueOnce(new Error('API Error'))
-    const { reportException } = useExceptionReporting()
+  it('handles API errors gracefully', async () => {
+    const apiError = new Error('API Error');
+    vi.mocked(api.post).mockRejectedValueOnce(apiError);
+    const { reportException } = useExceptionReporting();
 
-    await expect(reportException(mockExceptionData)).resolves.not.toThrow()
+    await expect(reportException(mockExceptionData)).resolves.not.toThrow();
 
-    expect(console.error).toHaveBeenCalledWith('Failed to report exception:', expect.any(Error))
-  })
-})
+    expect(consoleErrorSpy).toHaveBeenCalledWith('Failed to report exception:', apiError);
+  });
+});
